requirements.txt
torch>=2.0.0
transformers>=4.30.0
fair-esm>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.0
scipy>=1.10.0
tqdm>=4.65.0
umap-learn>=0.5.3

